LU-9679 general: add missing spaces to folded strings.
Many places in lustre fold a long string onto multiple lines,
usually at word breaks.  Sometimes the space between those words
got lost.
In a couple of places, a newline (n) rather than a space was lost.

This patch adds those spaces (and newlines) back in.

Where a space was added, the whole string is joined onto a
single line as this is current policy - encouraged by checkpatch.

In a couple of places neighbouring strings are also joined
into a single line, and some code has been re-indented to use
TABs.

Where the missing space was in a .diff file, the string hasn't
been joined into a line, as it doesn't seem worth the churn.

Test-Parameters: trivial
Signed-off-by: Mr NeilBrown <neilb@suse.de>
Change-Id: I6882bb957df566da0794f4ee85133dbf8c3debc1
Reviewed-on: https://review.whamcloud.com/36653
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: James Simmons <jsimmons@infradead.org>