LU-12937 utils: update wirecheck for new values
Update the wirecheck.c file to handle changes that were made directly
in the wiretest.c files.  This allows the wiretrst.c file to be
regenerated properly without losing checks for new constants, and
fixes issues with some #defines that were changed to named enums.

Code under CONFIG_FS_POSIX_ACL in utils/wirecheck.c does not build in
userspace if the <linux/posix_acl_xattr.h> header is not available.
It should be enough that we are checking this in ptlrpc/wirecheck.c
since struct posix_acl_xattr_entry and posix_acl_xattr_header come
from the kernel anyway, so a userspace check is mostly redundant.

Test-Parameters: trivial
Fixes: 1fd63fcb045c ("LU-12090 utils: lfs rmfid")
Fixes: b4375f5fc66c ("LU-11444 ptlrpc: Add increasing XIDs CONNECT2 flag")
Fixes: 19b2bc9bbc25 ("LU-6142 lustre: introduce CONFIG_LUSTRE_FS_POSIX_ACL")
Fixes: 3611352b699c ("LU-11285 mdt: improve IBITS lock definitions")
Fixes: 68635c3d9b31 ("LU-11963 osd: Add nonrotational flag to statfs")
Signed-off-by: Andreas Dilger <adilger@whamcloud.com>
Change-Id: I4945d6e2e8d7f4f98530e6dafd43bcf93735a5b1
Reviewed-on: https://review.whamcloud.com/36706
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Artem Blagodarenko <c17828@cray.com>
Reviewed-by: Arshad Hussain <arshad.super@gmail.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>