LU-13065 tests: clean up runtests
The runtests test suite needs to be cleaned up with the
following:

runtests used a verbose flag that initialized a ‘V’
variable, but setting the verbose option was removed.
The flag is still in the script.  The ‘V’ flag should be
replaced by the global VERBOSE flag set in
test-frameworks.sh.

Insert tabs at the beginning of each line of test 1

Remove ‘\’ after ||

The error() function does not need a second argument and may
be confusing when reading the error message.  Remove the
second argument from the calls to error()

Replace exit() calls with calls to error() where appropriate

Use lower case for all letters of local variables and
declare them 'local'

Test-Parameters: trivial testlist=runtests
Signed-off-by: James Nunez <jnunez@whamcloud.com>
Change-Id: I4e9f1eb467b004e5e2384dba4b7f047762304e8b
Reviewed-on: https://review.whamcloud.com/37016
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Wei Liu <sarah@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Emoly Liu <emoly@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>